memory : Removed ARM architecture dependencies in memory
Create interface level code regarding the memory.
and distinguish host(hyper mode) and guest in lpae.c
