--- platform-device/cortex_a15x2_arndale/config/cfg_platform.h
@@ -1,105 +1,0 @@
-/*
- *  BOARD param
- */
-#define CFG_BOARD_ARNDALE
-#define CFG_EXYNOS5250
-#define CFG_CNTFRQ          24000000
-#define CFG_UART2         0x12C00000
-#define CFG_UART1         0x12C10000
-#define CFG_UART0         0x12C20000
-
-/*
- *  SOC param
- */
-#define CFG_GIC_BASE_PA   0x10480000
-
-#define CFG_MACHINE_NUMBER 4274
-
-#define SZ_1                0x00000001
-#define SZ_2                0x00000002
-#define SZ_4                0x00000004
-#define SZ_8                0x00000008
-#define SZ_16               0x00000010
-#define SZ_32               0x00000020
-#define SZ_64               0x00000040
-#define SZ_128              0x00000080
-#define SZ_256              0x00000100
-#define SZ_512              0x00000200
-
-#define SZ_1K               0x00000400
-#define SZ_2K               0x00000800
-#define SZ_4K               0x00001000
-#define SZ_8K               0x00002000
-#define SZ_16K              0x00004000
-#define SZ_32K              0x00008000
-#define SZ_64K              0x00010000
-#define SZ_128K             0x00020000
-#define SZ_256K             0x00040000
-#define SZ_512K             0x00080000
-
-#define SZ_1M               0x00100000
-#define SZ_2M               0x00200000
-#define SZ_4M               0x00400000
-#define SZ_8M               0x00800000
-#define SZ_16M              0x01000000
-#define SZ_32M              0x02000000
-#define SZ_64M              0x04000000
-#define SZ_128M             0x08000000
-#define SZ_256M             0x10000000
-#define SZ_512M             0x20000000
-
-#define SZ_1G               0x40000000
-#define SZ_2G               0x80000000
-
-#define CFG_GUEST0_DEVICE_MEMORY \
-{ \
-{ "pl330.0", 0x11C10000, 0x11C10000, SZ_64K, LPAED_STAGE2_MEMATTR_DM }, \
-{ "pl330.1", 0x121A0000, 0x121A0000, SZ_64K, LPAED_STAGE2_MEMATTR_DM }, \
-{ "pl330.2", 0x121B0000, 0x121B0000, SZ_64K, LPAED_STAGE2_MEMATTR_DM }, \
-{ "uart.0", 0x12C00000, 0x12C00000, SZ_64K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "uart.1", 0x12C10000, 0x12C10000, SZ_64K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "uart.2", 0x12C20000, 0x12C20000, SZ_64K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "uart.3", 0x12C30000, 0x12C30000, SZ_64K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "chipid", 0x10000000, 0x10000000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },   \
-{ "syscon", 0x10050000, 0x10050000, SZ_64K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "timer", 0x12DD0000, 0x12DD0000, SZ_16K, LPAED_STAGE2_MEMATTR_DM },   \
-{ "wdt", 0x101D0000, 0x101D0000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },      \
-{ "sromc", 0x12250000, 0x12250000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },    \
-{ "hsphy", 0x12130000, 0x12130000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },    \
-{ "systimer", 0x101C0000, 0x101C0000, SZ_4K, LPAED_STAGE2_MEMATTR_DM }, \
-{ "sysram", 0x02020000, 0x02020000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },   \
-{ "cmu", 0x10010000, 0x10010000, 144 * SZ_1K, LPAED_STAGE2_MEMATTR_DM },\
-{ "pmu", 0x10040000, 0x10040000, SZ_64K, LPAED_STAGE2_MEMATTR_DM },     \
-{ "combiner", 0x10440000, 0x10440000, SZ_4K, LPAED_STAGE2_MEMATTR_DM }, \
-{ "gpio1", 0x11400000, 0x11400000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },    \
-{ "gpio2", 0x13400000, 0x13400000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },    \
-{ "gpio3", 0x10D10000, 0x10D10000, SZ_256, LPAED_STAGE2_MEMATTR_DM },   \
-{ "gpio4", 0x03860000, 0x03860000, SZ_256, LPAED_STAGE2_MEMATTR_DM },   \
-{ "audss", 0x03810000, 0x03810000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },    \
-{ "hsphy", 0x12130000, 0x12130000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },    \
-{ "ss_phy", 0x12100000, 0x12100000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },   \
-{ "sysram_ns", 0x0204F000, 0x0204F000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },\
-{ "ppmu_cpu", 0x10C60000, 0x10C60000, SZ_8K, LPAED_STAGE2_MEMATTR_DM }, \
-{ "ppmu_ddr_c", 0x10C40000, 0x10C40000, SZ_8K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "ppmu_ddr_r1", 0x10C50000, 0x10C50000, SZ_8K, LPAED_STAGE2_MEMATTR_DM }, \
-{ "ppmu_ddr_l", 0x10CB0000, 0x10CB0000, SZ_8K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "ppmu_right0_bus", 0x13660000, 0x13660000, SZ_8K, \
-    LPAED_STAGE2_MEMATTR_DM}, \
-{ "fimc_lite0", 0x13C00000, 0x13C00000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "fimc_lite1", 0x13C10000, 0x13C10000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "fimc_lite2", 0x13C90000, 0x13C90000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "mipi_csis0", 0x13C20000, 0x13C20000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "mipi_csis1", 0x13C30000, 0x13C30000, SZ_4K, LPAED_STAGE2_MEMATTR_DM },  \
-{ "gicc", CFG_GIC_BASE_PA | GIC_OFFSET_GICC, \
-    CFG_GIC_BASE_PA | GIC_OFFSET_GICVI, 0x2000, LPAED_STAGE2_MEMATTR_DM }, \
-{ 0, 0, 0, 0, 0 } \
-}
-
-#define CFG_GUEST1_DEVICE_MEMORY \
-{ \
-{ "uart", 0x12C10000, 0x12C20000, 0x1000, LPAED_STAGE2_MEMATTR_DM },       \
-{ "pwm_timer", 0x3FD10000, 0x12DD0000, 0x1000, LPAED_STAGE2_MEMATTR_DM },  \
-{ "gicc", CFG_GIC_BASE_PA | GIC_OFFSET_GICC, \
-    CFG_GIC_BASE_PA | GIC_OFFSET_GICVI, 0x2000, LPAED_STAGE2_MEMATTR_DM }, \
-{ 0, 0, 0, 0, 0 } \
-}
