--- platform-device/cortex_a15x2_arndale/config/memmap.cfg
@@ -1,7 +1,0 @@
-#define CFG_MEMMAP_PHYS_START      0x40000000
-#define CFG_MEMMAP_PHYS_SIZE       0x7FFFFFFF
-#define CFG_MEMMAP_PHYS_END        (CFG_MEMMAP_PHYS_START+CFG_MEMMAP_PHYS_SIZE)
-#define CFG_MEMMAP_MON_OFFSET      0xA0000000
-#define CFG_MEMMAP_GUEST_OFFSET    0x60000000
-#define CFG_MEMMAP_GUEST2_OFFSET   0x90000000
-
