+++ platform-device/cortex_a15x2_arndale/model.lds.S
@@ -12,7 +12,7 @@ OUTPUT_ARCH(arm)
 TARGET(binary)
 ENTRY(start)
 
-#include "config/memmap.cfg"
+#include "k-hypervisor-config.h"
 
 MON_SIZE    = 0x0F000000;
 MON_STACK	= CFG_MEMMAP_MON_OFFSET + MON_SIZE;
@@ -40,8 +40,8 @@ SECTIONS {
 
     /* Guest 1 */
     . = CFG_MEMMAP_GUEST_OFFSET;
-    guest_bin_start = .;
-    guest_bin_end = .;
+    _guest_bin_start = .;
+    _guest_bin_end = .;
 
     . = GUEST_STACK;
     guest_stacktop = .;
@@ -50,8 +50,8 @@ SECTIONS {
 
     /* Guest 2 */
     . = CFG_MEMMAP_GUEST2_OFFSET;
-    guest2_bin_start = .;
-    guest2_bin_end = .;
+    _guest2_bin_start = .;
+    _guest2_bin_end = .;
 
     . = GUEST2_STACK;
     guest2_stacktop = .;
